Removing iterators over Stormpath collections (these should page automatically); adding a decorator over account iterator
